Solved Database issue of parcel model. Not tested though

+++ app/src/main/java/com/scleroid/nemai/models/ParcelDao.java
@@ -19,7 +19,7 @@ public interface ParcelDao {
     List<Parcel> getAll();
 
     @Query("SELECT * FROM parcel where serialNo = :serialNo ")
-    Parcel findById(int serialNo);
+    Parcel findById(long serialNo);
 
     @Query("SELECT COUNT(*) from Parcel")
     int countParcels();
